pyproject.toml
[project]
name = "cfo-agent"
version = "0.1.0"
description = "NL2SQL financial BI agent for P&L data via Microsoft Teams"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-community>=0.3",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-anthropic>=0.3",
    "sqlalchemy>=2.0",
    "pyodbc>=5.0",
    "pydantic-settings>=2.0",
    "jinja2>=3.1",
    "botbuilder-core>=4.16",
    "aiohttp>=3.9",
    "structlog>=24.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "ollama: requires local Ollama server",
    "slow: takes more than 5 seconds",
]
testpaths = ["tests"]
